+++ djangae/urls.py
@@ -1,6 +1,7 @@
 from django.conf.urls import url, include
+import djangae.contrib.pipelines.urls
+import djangae.contrib.mapreduce.urls
 
-import djangae.contrib.mappers.urls
 from . import views
 
 
@@ -10,5 +11,6 @@ urlpatterns = [
     url(r'^warmup$', views.warmup),
     url(r'^queue/deferred/?$', views.deferred),
     url(r'^internalupload/$', views.internalupload, name='djangae_internal_upload_handler'),
-    url(r'^mapreduce/', include(djangae.contrib.mappers.urls))
+    url(r'^pipeline/', include(djangae.contrib.pipelines.urls)),
+    url(r'^mapreduce/', include(djangae.contrib.mapreduce.urls)),
 ]