+++ testapp/testapp/settings.py
@@ -44,7 +44,8 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'django.contrib.humanize',
-    'djangae.contrib.mappers',
+    'djangae.contrib.pipelines',
+    'djangae.contrib.mapreduce',
     'djangae.contrib.pagination',
     'djangae.contrib.uniquetool',
     'testapp'